Jules was unable to complete the task in time. Please review the work done so far and provide feedback for Jules to continue.
