+++ src/app/layout.tsx
@@ -1,10 +1,10 @@
 import type { Metadata } from "next";
 import "./globals.css";
-import { SearchDropdownComponent } from "@/components/search-dropdown";
+import { SearchDropdownComponent } from "@/features/search/components/search-dropdown";
 import { MenuIcon } from "lucide-react";
 import { Suspense } from "react";
-import { Cart } from "@/components/cart";
-import { AuthServer } from "./auth.server";
+import CartIcon from "@/features/cart/components/CartIcon"; // Updated import
+import { AuthServer } from "@/features/auth/auth.server";
 import { Link } from "@/components/ui/link";
 import { Analytics } from "@vercel/analytics/react";
 import { Toaster } from "sonner";
@@ -71,7 +71,7 @@ export default async function RootLayout({
                         ORDER
                       </Link>
                       <Suspense>
-                        <Cart />
+                        <CartIcon /> {/* Updated usage */}
                       </Suspense>
                     </div>
                     <Link
