+++ src/features/auth/lib/queries.ts
@@ -1,0 +1,38 @@
+import { cookies } from "next/headers";
+import { verifyToken } from "@/features/auth/lib/session";
+import { users } from "@/db/schema";
+import { db } from "@/db";
+import { eq, and } from "drizzle-orm";
+import { unstable_cache } from "@/lib/unstable-cache";
+
+export async function getUser() {
+  const sessionCookie = (await cookies()).get("session");
+  if (!sessionCookie || !sessionCookie.value) {
+    return null;
+  }
+
+  const sessionData = await verifyToken(sessionCookie.value);
+  if (
+    !sessionData ||
+    !sessionData.user ||
+    typeof sessionData.user.id !== "number"
+  ) {
+    return null;
+  }
+
+  if (new Date(sessionData.expires) < new Date()) {
+    return null;
+  }
+
+  const user = await db
+    .select()
+    .from(users)
+    .where(and(eq(users.id, sessionData.user.id)))
+    .limit(1);
+
+  if (user.length === 0) {
+    return null;
+  }
+
+  return user[0];
+}
