+++ src/features/orders/components/TotalCost.tsx
@@ -1,0 +1,12 @@
+import { getCartItems } from "@/features/cart/lib/getCartItems";
+
+export async function TotalCost() {
+  const cart = await getCartItems();
+
+  const totalCost = cart.reduce(
+    (acc, item) => acc + item.quantity * Number(item.price),
+    0,
+  );
+
+  return <span> ${totalCost.toFixed(2)}</span>;
+}
