+++ src/features/plp/components/CategoryItem.tsx
