+++ src/features/search/lib/queries.ts
@@ -1,0 +1,69 @@
+import { db } from "@/db";
+import {
+  products,
+  categories,
+  subcategories,
+  subcollections,
+} from "@/db/schema";
+import { sql } from "drizzle-orm";
+import { unstable_cache } from "@/lib/unstable-cache";
+
+export const getSearchResults = unstable_cache(
+  async (searchTerm: string) => {
+    let results;
+
+    // do we really need to do this hybrid search pattern?
+
+    if (searchTerm.length <= 2) {
+      // If the search term is short (e.g., "W"), use ILIKE for prefix matching
+      results = await db
+        .select()
+        .from(products)
+        .where(sql`${products.name} ILIKE ${searchTerm + "%"}`) // Prefix match
+        .limit(5)
+        .innerJoin(
+          subcategories,
+          sql`${products.subcategory_slug} = ${subcategories.slug}`,
+        )
+        .innerJoin(
+          subcollections,
+          sql`${subcategories.subcollection_id} = ${subcollections.id}`,
+        )
+        .innerJoin(
+          categories,
+          sql`${subcollections.category_slug} = ${categories.slug}`,
+        );
+    } else {
+      // For longer search terms, use full-text search with tsquery
+      const formattedSearchTerm = searchTerm
+        .split(" ")
+        .filter((term) => term.trim() !== "") // Filter out empty terms
+        .map((term) => `${term}:*`)
+        .join(" & ");
+
+      results = await db
+        .select()
+        .from(products)
+        .where(
+          sql`to_tsvector('english', ${products.name}) @@ to_tsquery('english', ${formattedSearchTerm})`,
+        )
+        .limit(5)
+        .innerJoin(
+          subcategories,
+          sql`${products.subcategory_slug} = ${subcategories.slug}`,
+        )
+        .innerJoin(
+          subcollections,
+          sql`${subcategories.subcollection_id} = ${subcollections.id}`,
+        )
+        .innerJoin(
+          categories,
+          sql`${subcollections.category_slug} = ${categories.slug}`,
+        );
+    }
+
+    return results;
+  },
+  ["search-results"],
+  { revalidate: 60 * 60 * 2 }, // two hours
+);
